RENAME : separate entity and dto
