+++ Effi-BE-fin/src/main/java/com/example/effi/domain/DTO/CategoryDTO.java
@@ -1,5 +1,6 @@
-package com.example.effi.domain;
+package com.example.effi.domain.DTO;
 
+import com.example.effi.domain.Entitiy.Category;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
