+++ Effi-BE-fin/src/main/java/com/example/effi/domain/DTO/GroupEmpDTO.java
@@ -1,5 +1,8 @@
-package com.example.effi.domain;
+package com.example.effi.domain.DTO;
 
+import com.example.effi.domain.Entitiy.Employee;
+import com.example.effi.domain.Entitiy.Group;
+import com.example.effi.domain.Entitiy.GroupEmp;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
