+++ Effi-BE-fin/src/main/java/com/example/effi/domain/DTO/RoutineDTO.java
@@ -1,5 +1,6 @@
-package com.example.effi.domain;
+package com.example.effi.domain.DTO;
 
+import com.example.effi.domain.Entitiy.Routine;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
