+++ Effi-BE-fin/src/main/java/com/example/effi/domain/DTO/ScheduleDTO.java
@@ -1,5 +1,8 @@
-package com.example.effi.domain;
+package com.example.effi.domain.DTO;
 
+import com.example.effi.domain.Entitiy.Category;
+import com.example.effi.domain.Entitiy.Routine;
+import com.example.effi.domain.Entitiy.Schedule;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
