+++ Effi-BE-fin/src/main/java/com/example/effi/domain/DTO/TagScheduleDTO.java
@@ -1,5 +1,8 @@
-package com.example.effi.domain;
+package com.example.effi.domain.DTO;
 
+import com.example.effi.domain.Entitiy.Schedule;
+import com.example.effi.domain.Entitiy.Tag;
+import com.example.effi.domain.Entitiy.TagSchedule;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
