+++ Effi-BE-fin/src/main/java/com/example/effi/domain/DTO/TimezoneDTO.java
@@ -1,5 +1,6 @@
-package com.example.effi.domain;
+package com.example.effi.domain.DTO;
 
+import com.example.effi.domain.Entitiy.Timezone;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
