+++ Effi-BE-fin/src/main/java/com/example/effi/domain/Entitiy/Employee.java
@@ -1,5 +1,6 @@
-package com.example.effi.domain;
+package com.example.effi.domain.Entitiy;
 
+import com.example.effi.domain.Entitiy.Dept;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
