+++ Effi-BE-fin/src/main/java/com/example/effi/domain/Entitiy/Group.java
@@ -1,5 +1,6 @@
-package com.example.effi.domain;
+package com.example.effi.domain.Entitiy;
 
+import com.example.effi.domain.Entitiy.Category;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
