+++ Effi-BE-fin/src/main/java/com/example/effi/domain/Entitiy/Schedule.java
@@ -1,5 +1,7 @@
-package com.example.effi.domain;
+package com.example.effi.domain.Entitiy;
 
+import com.example.effi.domain.Entitiy.Category;
+import com.example.effi.domain.Entitiy.Routine;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
