+++ Effi-BE-fin/src/main/java/com/example/effi/repository/EmployeeRepository.java
@@ -3,7 +3,7 @@ package com.example.effi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.effi.domain.Employee;
+import com.example.effi.domain.Entitiy.Employee;
 
 @Repository
 public interface EmployeeRepository extends JpaRepository<Employee, Long>{
