+++ Effi-BE-fin/src/main/java/com/example/effi/service/EmployeeService.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.effi.config.TokenProvider;
-import com.example.effi.domain.Employee;
-import com.example.effi.domain.SignInRequest;
-import com.example.effi.domain.SignInResponse;
+import com.example.effi.domain.Entitiy.Employee;
+import com.example.effi.domain.DTO.SignInRequest;
+import com.example.effi.domain.DTO.SignInResponse;
 import com.example.effi.repository.EmployeeRepository;
 
 import lombok.RequiredArgsConstructor;
